redpanda: split debug endpoints into separate translation unit
* Renamed logger to adminlog
* Moved some helpers into class as static methods
